Changed auth flow for Reddit authorization
- Since the Tunestumbler API cannot redirect the Tunestumbler website
to an external url, the API needs to return the authorization url
itself, that is generated server side, to the client. Then the client
will send the state and code to the API, which the API will use to
generate a Reddit authentication token. This token will be saved to
the user’s data in the database, and then the user will be able to
make Reddit requests just with the Tunestumbler token.
- Generate authorization url to send to the client to allow them to
authenticate with Reddit
- Expose Tunestumbler and Reddit authentication tokens and token
lifetimes in response header so that the client can easily use and
store them
- Expose userId in response header
- Save reddit token lifetime to user data and expose this in response
header so that the client can use this to reauthenticate and generate
a new Reddit token
- Reduce Tunestumbler token lifetime from 1 year to 3 weeks
- Increase minimum password length to 8 characters
- Changed the auth response model to reflect the above changes