Removed database access for subreddit, multireddit, and aggregate services and refactored functionality into separate helper classes
- Removed `SubredditEntity`, `MultiredditEntity`, `AggregateEntity`,
`SubredditRepository`, `MultiredditRepository`, and
`AggregateRepository` since the API now handles the Reddit subreddit
and multireddit responses directly, and returns the combined
aggregate result
- Removed API methods for `createAggregateByUserId`,
`updateAggregateByUserId`, `updateMultireddits`,
`getMultiredditsByUserId`, `updateSubreddits`,
`getSubredditsByUserId`. These methods fetch and store subreddits and
 multireddits in the database. However, this is no longer necessary.
Since the Reddit API has no webhooks, there is no way for the
Tunestumbler API to know when the user’s Reddit account is subscribed
 or unsubscribed to new subreddits, and therefore, the
Tunestumbler API will always need to ask the Reddit API for the
current subscribed subreddits whenever they need to be fetched to
create new filters.
- To bypass having to save subreddits to the database, the new
Tunestumbler API methods, `fetchSubreddits` and `fetchMultireddits`,
will be called within `getAggregateByUserId` to form an aggregate of
subscribed subreddits, which will then be sent back with the
HTTP response.
- Method functionality is also broken down into separate public
helper methods within the `SubredditHelpers` and `MultiredditHelpers`
 class files
- Removed `id`, `isSubredditAdded`, `startId`, `lastModified` fields
from the `AggregateDTO`. These fields were added along with the
records in the database for all of the user’s subreddits. Now that
the subreddits will only be passed back to the calling HTTP response,
 it makes sense to remove them from the DTO.
- Removed fields from the `SubredditDTO` and `MultiredditDTO` for similar
 reasons to removing fields from the `AggregateDTO`.
- Removed unused endpoints from `SubredditController`,
`MultiredditController`, and `AggregateController`