Normalized filters tables and reflected changes in filters service
- Filters data needs to be normalized so that the domain and keyword
filters are organized into their own tables
- Before normalization, the filters were structured such that there
would be multiple subreddit filters, and each may contain only one of
 each domain and keyword filter. This resulted in overlapping,
duplicated filtered Reddit results
- Normalizing the filters eliminates filtered result duplication,
and restructures each subreddit filter, such that each filter will
have its own list of hidden domains, hidden title keywords, visible
domains, and visible title keyword filters
- But normalizing the filters also means that the requests and
 responses for creating and updating filters needs to change
- In this feature branch, I decided to structure the `POST` and `PUT`
 requests such that the API consumer simply needs to tell the API how
 they expect their filters to look. This means that the body of the
requests must contain the list of filters, along with the complete
lists of subfilters (ie: ExcludedDomains, SelectedKeywords), that the
 consumer wants to create or update, with the exception of delete.
- To delete filters or subfilters, there will be dedicated endpoints
since it is simple to delete a list of ids
- This also lets the tunestumbler website to restructure the filters
UI so that it looks much more organized and less busy

- Renamed `HideByKeyword`, `ShowByKeyword`, `HideByDomain`,
`ShowByDomain` subfilters to `ExcludedKeywords`, `SelectedDomains`,
`ExcludedDomains`, `SelectedKeywords` so that it is more clear what
the parameters represent
- `Domain` and `keyword` filters have a `@OneToMany` unidirectional
association where `FilterEntity` is the association owner, so that
 when a filter is deleted, its domain and keyword filters are also
 deleted
- `FetchType.LAZY` is used to lazy load  (ie: data is fetched only
when called), and the `JoinColumn` annotation is used in the owner
table to specify the column that is referenced in the referenced
table
- The `parent_filters_id` column in domain and filter tables is used
to reference the filter owner of the unidirectional relationship
- Remove `isActive` fields from all filter entities since this field
was used to denote the filters that should be batch deleted manually.
 This will be replaced with separate `delete` endpoints to delete
inactive filters
- Updated the filters repository query since the `is_active` property
 is no longer used
- Updated `FiltersDTO` to remove the `isActive` field
- Modified `FiltersRepository` query so that the results are ordered
by ascending id since order is not guaranteed when querying
- Updated filters service methods to reflect the changes in the
filters database tables
- Moved filters service functionality into separate helper methods so
 that the code is more readable, maintainable, and testable
- Updated filters request and response models to reflect database
changes

- Added `Filter` repository method to query all owned filters by a
list of `filterId`s and` userId`
- Removed unused `/{userId}` endpoints
- Added repository methods for subfilters
- Added dedicated endpoints and service methods for deleting multiple
 filters and subfilters at a time
- MapStruct was used to partially generate mapping classes which
contain methods for mapping between filter and subfilter data types.
This is because `Filters` subfilters are now stored in collections,
and the mapping has become complicated
- Altered `createFilters` and `updateFilters` service methods to use
the mapper methods to convert between request objects, DTOs,
Entities, and response objects
- When mapping updated filters, a hashmap is used to more easily map
updated filter data to outdated filter data

- Swapped out modelmapper usage for custom filter mapper, since it
is less verbose
- Added repository method to find the set of subreddits by userId, so
 that duplicate filters for the same subreddit are not created.
Duplicate updates for the same subreddit are not prevented since a
single mistake could discard all of the end users changes, and I
think it would be a better experience to instead overwrite filters
with the latest filter fields
- An error is thrown if the consumer tries to create new filters for
an already existing subreddit filter
- Created separate request objects and mappers for creating and
updating filters, because creating filters does not require a
`filterId`, and so create objects should have different validation
checks compared to update objects
- Removed unused endpoints that require `userId` in the path